Fix bug

+++ app/src/main/java/com/fongmi/android/tv/bean/Track.java
@@ -1,5 +1,7 @@
 package com.fongmi.android.tv.bean;
 
+import android.text.TextUtils;
+
 import androidx.annotation.NonNull;
 import androidx.room.Entity;
 import androidx.room.Index;
@@ -7,6 +9,7 @@ import androidx.room.PrimaryKey;
 
 import com.fongmi.android.tv.db.AppDatabase;
 
+import java.util.Collections;
 import java.util.List;
 
 @Entity(indices = @Index(value = {"key", "type"}, unique = true))
@@ -97,14 +100,16 @@ public class Track {
     }
 
     public void save() {
+        if (TextUtils.isEmpty(getKey())) return;
         AppDatabase.get().getTrackDao().insert(this);
     }
 
     public static List<Track> find(String key) {
-        return AppDatabase.get().getTrackDao().find(key);
+        return TextUtils.isEmpty(key) ? Collections.emptyList() : AppDatabase.get().getTrackDao().find(key);
     }
 
     public static void delete(String key) {
+        if (TextUtils.isEmpty(key)) return;
         AppDatabase.get().getTrackDao().delete(key);
     }
 }
